restructure the clients into separate modules
